add configurable "limit_from" value
limit_from will override the "from" address that is passed to the milter
plugin query() method.  This can be any Sendmail macro.   For example,
if you wish to limit based on the SMTP authenticated user, you can use

    limit_from = {auth_authen}

in the global section of the config file.  The default continues to be
the envelope sender, and will fall back to the envelope sender if the
configured macro is not set

Fixes #7